+++ scripts/macros/monsterFeatures/elderBrain/psychicLink.js
@@ -31,6 +31,11 @@ async function item({speaker, actor, token, character, item, args}) {
                     'script': chris.functionToString(effectMacro)
                 }
             },
+            'chris-premades': {
+                'vae': {
+                    'button': featureData.name
+                }
+            }
         }
     };
     let updates = {