+++ scripts/macros/monsterFeatures/fireElemental/fireForm.js
@@ -50,6 +50,11 @@ export async function item({speaker, actor, token, character, item, args}) {
                     'script': chris.functionToString(effectMacro)
                 }
             },
+            'chris-premades': {
+                'vae': {
+                    'button': featureData.name
+                }
+            }
         }
     };
     let updates = {