+++ scripts/macros/spells/huntersMark.js
@@ -67,6 +67,11 @@ async function item({speaker, actor, token, character, item, args}) {
                 'onDelete': {
                     'script': chris.functionToString(effectMacro)
                 }
+            },
+            'chris-premades': {
+                'vae': {
+                    'button': featureData.name
+                }
             }
         }
     }