+++ scripts/macros/spells/vampiricTouch.js
@@ -34,6 +34,11 @@ async function vampiricTouchItem({speaker, actor, token, character, item, args})
 					'script': chris.functionToString(effectMacro)
 				}
 			},
+			'chris-premades': {
+				'vae': {
+					'button': featureData.name
+				}
+			}
 		}
 	};
 	let updates = {